refactor: split metrics tests into metrics/ directory, rename stale CRC/checkpoint scenarios
