+++ kernel/tests/metrics/snapshot_load.rs
@@ -1,104 +1,26 @@
-//! Integration tests that verify [`CountingReporter`] metrics across different table
-//! states and API call patterns.
+//! Snapshot-loading and on-demand API metrics tests.
 //!
-//! All tests use only public kernel APIs (`create_table`, `Transaction`, `write_parquet`,
-//! `log_compaction_writer`, `scan.execute()`) to ensure that the metrics surface is
-//! verified through the same code paths a real engine connector would exercise.
-//!
-//! Scenarios covered:
-//! - Delta-only snapshot (baseline)
-//! - V1 checkpoint with/without tail commits
-//! - Log compaction
-//! - CRC fast-path (P+M bypasses JSON replay)
-//! - Stale CRC (CRC older than latest version -- tail replay + CRC fallback)
-//! - Stale checkpoint (checkpoint + multiple tail commits)
-//! - On-demand domain metadata query (`get_domain_metadata`)
-//! - `snapshot.transaction()` on a clustered table (loads clustering metadata)
-//! - Scan data-file reads (`scan.execute()` contributes parquet I/O)
-//! - Incremental snapshot update (`Snapshot::builder_from` replays only new commits)
+//! Covers all major log-segment shapes (delta-only, V1 checkpoint, log compaction, CRC
+//! fast-path, CRC at prior version, checkpoint with tail commits) plus on-demand API calls
+//! (`get_domain_metadata`, `scan.execute()`) that incur additional I/O after a snapshot is
+//! already built.
 
+use super::{measuring_engine, setup_table_with_v1_checkpoint, simple_schema};
 use std::path::PathBuf;
 use std::sync::Arc;
 
 use delta_kernel::arrow::array::Int32Array;
 use delta_kernel::committer::FileSystemCommitter;
-use delta_kernel::engine::default::executor::tokio::TokioMultiThreadExecutor;
-use delta_kernel::engine::default::{DefaultEngine, DefaultEngineBuilder};
+use delta_kernel::engine::default::DefaultEngineBuilder;
 use delta_kernel::engine::to_json_bytes;
 use delta_kernel::object_store::local::LocalFileSystem;
 use delta_kernel::object_store::memory::InMemory;
 use delta_kernel::object_store::path::Path;
 use delta_kernel::object_store::ObjectStore as _;
-use delta_kernel::schema::{DataType, StructField, StructType};
 use delta_kernel::transaction::create_table::create_table;
 use delta_kernel::{DeltaResult, Engine, Snapshot};
-
-use test_utils::{insert_data, test_table_setup_mt, CountingReporter};
+use test_utils::{insert_data, test_table_setup_mt};
 use url::Url;
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/// Build a `DefaultEngine` + `CountingReporter` backed by `store`, for use in the
-/// *measurement* phase (building a snapshot and asserting metric counters).
-fn measuring_engine(
-    store: Arc<dyn delta_kernel::object_store::ObjectStore>,
-) -> (
-    DefaultEngine<delta_kernel::engine::default::executor::tokio::TokioBackgroundExecutor>,
-    Arc<CountingReporter>,
-) {
-    let reporter = Arc::new(CountingReporter::default());
-    let engine = DefaultEngineBuilder::new(store)
-        .with_metrics_reporter(reporter.clone())
-        .build();
-    (engine, reporter)
-}
-
-/// Build a minimal single-column INTEGER schema for test tables.
-fn simple_schema() -> Arc<StructType> {
-    Arc::new(
-        StructType::try_new(vec![StructField::nullable("id", DataType::INTEGER)])
-            .expect("valid schema"),
-    )
-}
-
-/// Create a table at v0, insert one row, and write a v1 parquet checkpoint.
-///
-/// Returns `(table_url, setup_engine, _temp_dir)` where `_temp_dir` must be kept
-/// alive for the duration of the test to prevent early cleanup. Callers can add
-/// further commits via `insert_data` or directly build the measuring snapshot.
-///
-/// Uses `TokioMultiThreadExecutor` because `snapshot.checkpoint()` issues nested
-/// `block_on` calls; `TokioBackgroundExecutor` deadlocks in that pattern.
-async fn setup_table_with_v1_checkpoint() -> DeltaResult<(
-    Url,
-    Arc<DefaultEngine<TokioMultiThreadExecutor>>,
-    tempfile::TempDir,
-)> {
-    let (temp_dir, table_path, setup_engine) = test_table_setup_mt()?;
-    let table_url = delta_kernel::try_parse_uri(&table_path)?;
-
-    let _ = create_table(&table_path, simple_schema(), "Test/1.0")
-        .build(setup_engine.as_ref(), Box::new(FileSystemCommitter::new()))?
-        .commit(setup_engine.as_ref())?;
-
-    let snap0 = Snapshot::builder_for(table_url.clone()).build(setup_engine.as_ref())?;
-    let committed = insert_data(
-        snap0,
-        &setup_engine,
-        vec![Arc::new(Int32Array::from(vec![1]))],
-    )
-    .await?
-    .unwrap_committed();
-    committed
-        .post_commit_snapshot()
-        .expect("post-commit snapshot")
-        .clone()
-        .checkpoint(setup_engine.as_ref())?;
-
-    Ok((table_url, setup_engine, temp_dir))
-}
 
 // ---------------------------------------------------------------------------
 // Scenario 1: delta-only (2 commits, no checkpoint, no compaction)
@@ -319,7 +241,7 @@ async fn snapshot_with_crc_at_target_version_skips_json_replay() -> DeltaResult<
 }
 
 // ---------------------------------------------------------------------------
-// Scenario 6: stale CRC (CRC behind latest version)
+// Scenario 6: CRC at a prior version (CRC exists but is older than latest)
 // ---------------------------------------------------------------------------
 
 /// When a CRC file exists at an older version than the snapshot, the kernel takes the
@@ -327,11 +249,14 @@ async fn snapshot_with_crc_at_target_version_skips_json_replay() -> DeltaResult<
 /// looking for Protocol/Metadata changes, then falls back to the CRC when none are
 /// found. This produces JSON reads for the tail commits AND a storage read for the CRC.
 ///
+/// This is a normal, expected table state -- having a CRC from a previous version is
+/// not an error or degraded condition.
+///
 /// `write_checksum` requires stats from a post-commit snapshot (not a freshly built
 /// one), so this test uses `create_table` commit -> `post_commit_snapshot` ->
 /// `write_checksum` to write the CRC at v0.
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn stale_crc_replays_tail_commits_and_loads_crc_as_fallback() -> DeltaResult<()> {
+async fn crc_at_prior_version_triggers_tail_replay_then_falls_back_to_crc() -> DeltaResult<()> {
     let (_temp_dir, table_path, setup_engine) = test_table_setup_mt()?;
     let table_url = delta_kernel::try_parse_uri(&table_path)?;
 
@@ -358,7 +283,7 @@ async fn stale_crc_replays_tail_commits_and_loads_crc_as_fallback() -> DeltaResu
         .await?;
     }
 
-    // Measurement: build snapshot at latest (v2); CRC is stale (at v0, not v2)
+    // Measurement: build snapshot at latest (v2); CRC is at v0 (two versions behind)
     let (measure_engine, reporter) = measuring_engine(Arc::new(LocalFileSystem::new()));
     let _snap = Snapshot::builder_for(table_url).build(&measure_engine)?;
 
@@ -366,7 +291,7 @@ async fn stale_crc_replays_tail_commits_and_loads_crc_as_fallback() -> DeltaResu
     assert_eq!(reporter.log_segment_loads.get(), 1);
     assert_eq!(reporter.commit_files.get(), 3); // v0, v1, v2
 
-    // Tail replay: commits v1 and v2 (after the stale CRC at v0) are replayed via JSON
+    // Tail replay: commits v1 and v2 (after the CRC at v0) are replayed via JSON
     assert_eq!(reporter.json_read_calls.get(), 1);
     assert_eq!(reporter.json_files_read.get(), 2);
     // CRC is loaded from storage as a fallback (P+M not found in the tail commits)
@@ -376,14 +301,15 @@ async fn stale_crc_replays_tail_commits_and_loads_crc_as_fallback() -> DeltaResu
 }
 
 // ---------------------------------------------------------------------------
-// Scenario 7: stale checkpoint (checkpoint with many tail commits)
+// Scenario 7: checkpoint behind latest version (with multiple tail commits)
 // ---------------------------------------------------------------------------
 
 /// A checkpoint that is multiple versions behind the latest forces a longer tail replay.
-/// Checkpoint at v1 plus 3 additional commits verifies that all tail commits are read
-/// in a single JSON call and both byte counters are non-zero.
+/// Checkpoint at v1 plus 3 additional commits (v2, v3, v4) verifies that all tail commits
+/// are read in a single JSON call and both byte counters are non-zero. The specific counts
+/// here reflect this table's setup: checkpoint at v1, three tail commits.
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn stale_checkpoint_with_multiple_tail_commits_emits_expected_metrics() -> DeltaResult<()> {
+async fn checkpoint_with_multiple_tail_commits_emits_expected_metrics() -> DeltaResult<()> {
     let (table_url, setup_engine, _temp_dir) = setup_table_with_v1_checkpoint().await?;
 
     // commits 2, 3, 4: insert more data after the checkpoint
@@ -402,7 +328,7 @@ async fn stale_checkpoint_with_multiple_tail_commits_emits_expected_metrics() ->
 
     assert_eq!(reporter.snapshot_completions.get(), 1);
     assert_eq!(reporter.log_segment_loads.get(), 1);
-    // Checkpoint at v1 filters out v0 and v1; tail is v2, v3, v4
+    // Checkpoint at v1 -- listing starts from v2; tail is v2, v3, v4
     assert_eq!(reporter.commit_files.get(), 3);
     assert_eq!(reporter.checkpoint_files.get(), 1);
     // All 3 tail commits read in a single JSON call; checkpoint in a single Parquet call
@@ -419,11 +345,13 @@ async fn stale_checkpoint_with_multiple_tail_commits_emits_expected_metrics() ->
 // Scenario 8: on-demand domain metadata query incurs additional log replay
 // ---------------------------------------------------------------------------
 
-/// `snapshot.get_domain_metadata()` always performs a full log replay (no CRC at
-/// target version here). Calling it after a snapshot is already built generates a
-/// second round of JSON reads, demonstrating that on-demand metadata queries carry
-/// their own I/O cost. `get_clustering_columns` uses the same internal path, so
-/// this test covers both APIs.
+/// `snapshot.get_domain_metadata()` always performs a full log replay when no CRC is
+/// present at the target version. Calling it after a snapshot is already built generates
+/// a second round of JSON reads, demonstrating that on-demand metadata queries carry
+/// their own I/O cost.
+///
+/// The specific count (`json_read_calls = 1`) reflects this test's table: 2 commits, no
+/// checkpoint, no CRC. Tables with different log structures will produce different counts.
 #[tokio::test]
 async fn get_domain_metadata_incurs_additional_log_replay() -> DeltaResult<()> {
     let store = Arc::new(InMemory::new());
@@ -462,7 +390,7 @@ async fn get_domain_metadata_incurs_additional_log_replay() -> DeltaResult<()> {
 }
 
 // ---------------------------------------------------------------------------
-// Scenario 9: snapshot.transaction() on a clustered table loads clustering metadata
+// Scenario 9: snapshot.transaction() on a clustered table
 // ---------------------------------------------------------------------------
 // NOTE: This scenario is verified via clustering_e2e.rs tests because the
 // `clustered-table` Rust feature is not exposed to the integration test binary.
@@ -524,61 +452,6 @@ async fn scan_execute_contributes_parquet_data_file_reads() -> DeltaResult<()> {
     assert!(reporter.parquet_bytes_read.get() > 0);
     // scan.execute() does its own log replay for Add/Remove scan metadata
     assert_eq!(reporter.json_read_calls.get(), 1);
-
-    Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Scenario 11: incremental snapshot update replays only new tail commits
-// ---------------------------------------------------------------------------
-
-/// `Snapshot::builder_from(existing)` starts from an existing snapshot and replays
-/// only the commits that arrived after it. The JSON reads reflect only the new tail
-/// commit, not the full history -- demonstrating that incremental updates are cheaper.
-///
-/// Note: `LogSegmentLoaded` is emitted with the net-new commit count only (the new
-/// commits after the existing snapshot version), not the full log history.
-#[tokio::test]
-async fn incremental_snapshot_update_replays_only_new_commits() -> DeltaResult<()> {
-    let store = Arc::new(InMemory::new());
-    let table_url = Url::parse("memory:///").unwrap();
-    let setup_engine = Arc::new(DefaultEngineBuilder::new(store.clone() as Arc<_>).build());
-
-    let _ = create_table("memory:///", simple_schema(), "Test/1.0")
-        .build(setup_engine.as_ref(), Box::new(FileSystemCommitter::new()))?
-        .commit(setup_engine.as_ref())?;
-    let snap0 = Snapshot::builder_for(table_url.clone()).build(setup_engine.as_ref())?;
-    let _ = insert_data(
-        snap0,
-        &setup_engine,
-        vec![Arc::new(Int32Array::from(vec![1]))],
-    )
-    .await?;
-
-    // Capture snapshot at v1 (no reporter needed for setup)
-    let existing_snap = Snapshot::builder_for(table_url.clone()).build(setup_engine.as_ref())?;
-
-    // Add one more commit (v2) after the captured snapshot
-    let _ = insert_data(
-        Snapshot::builder_for(table_url.clone()).build(setup_engine.as_ref())?,
-        &setup_engine,
-        vec![Arc::new(Int32Array::from(vec![2]))],
-    )
-    .await?;
-
-    // Incremental update from existing_snap (at v1) with a reporter
-    let (engine, reporter) = measuring_engine(store);
-    let _updated = Snapshot::builder_from(existing_snap).build(&engine)?;
-
-    assert_eq!(reporter.snapshot_completions.get(), 1);
-    // LogSegmentLoaded is emitted with the net-new commit count (1 new commit: v2)
-    assert_eq!(reporter.log_segment_loads.get(), 1);
-    assert_eq!(reporter.commit_files.get(), 1);
-
-    // JSON handler reads only commit v2, not the full 3-commit history
-    assert_eq!(reporter.json_read_calls.get(), 1);
-    assert_eq!(reporter.json_files_read.get(), 1);
-    assert!(reporter.list_calls.get() >= 1);
 
     Ok(())
 }
